requirements.txt
appdirs==1.4.4
packaging==25.0
pillow==11.3.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
PySide6==6.9.2
